+++ BlueprintUICommonControls/Tests/Sources/PerformancePlayground.swift
@@ -51,6 +51,102 @@ class PerformancePlayground: XCTestCase {
         }
     }
 
+    func test_rows_with_one_flexible_element_simple() {
+
+        let row = Row(alignment: .fill, underflow: .growUniformly, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+            Box(backgroundColor: .red)
+                .constrainedTo(width: 80, height: 80)
+                .stackLayoutChild(priority: .fixed)
+
+            Column(alignment: .fill, underflow: .justifyToStart, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+                let rows: [String] = [
+                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
+                ]
+
+                for rowText in rows {
+                    Row(alignment: .fill, underflow: .growUniformly, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+                        Box(backgroundColor: .blue)
+                            .constrainedTo(width: 40, height: 40)
+                            .stackLayoutChild(priority: .fixed)
+
+                        Label(text: rowText)
+                            .stackLayoutChild(priority: .flexible)
+                    }
+                    .stackLayoutChild(priority: .fixed)
+                }
+            }
+            .stackLayoutChild(priority: .flexible)
+        }
+
+        let content = Column(alignment: .fill) {
+            for _ in 1...10 {
+                row
+            }
+        }
+
+        let size = content.content.measure(in: .init(width: 300))
+
+        let view = BlueprintView()
+        view.frame.size = CGSize(width: 300.0, height: size.height)
+
+        view.element = content
+        view.layoutIfNeeded()
+    }
+
+    func test_rows_with_one_flexible_element() {
+
+        let row = Row(alignment: .fill, underflow: .growUniformly, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+            Box(backgroundColor: .red)
+                .constrainedTo(width: 80, height: 80)
+                .stackLayoutChild(priority: .fixed)
+
+            Column(alignment: .fill, underflow: .justifyToStart, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+                let rows: [String] = [
+                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
+                    "Fusce dignissim vitae leo sed pretium.",
+                    "Morbi molestie, nisl eget faucibus bibendum, magna orci congue ipsum.",
+                    "Eu tincidunt orci nunc ac nisl.",
+                    "Aenean facilisis nulla vitae nibh suscipit, id placerat est lacinia.",
+                ]
+
+                for rowText in rows {
+                    Row(alignment: .fill, underflow: .growUniformly, overflow: .condenseUniformly, minimumSpacing: 10) {
+
+                        Box(backgroundColor: .blue)
+                            .constrainedTo(width: 40, height: 40)
+                            .stackLayoutChild(priority: .fixed)
+
+                        Label(text: rowText)
+                            .stackLayoutChild(priority: .flexible)
+                    }
+                    .stackLayoutChild(priority: .fixed)
+                }
+            }
+            .stackLayoutChild(priority: .flexible)
+        }
+
+        let allRows = Column(alignment: .fill) {
+            for _ in 1...100 {
+                row
+            }
+        }
+
+        let size = allRows.content.measure(in: .init(width: 300))
+
+        let view = BlueprintView()
+        view.frame.size = CGSize(width: 300.0, height: size.height)
+
+        determineAverage(for: 2.0) {
+            view.element = allRows
+            view.layoutIfNeeded()
+        }
+    }
+
     func test_deep_element_hierarchy() {
         let elements = lipsumStrings.map {
             Label(text: $0)