+++ Package.swift
@@ -25,6 +25,8 @@ let package = Package(
         .target(
             name: "BlueprintUI",
             path: "BlueprintUI/Sources"
+            // Enable this setting to allow running tests in release mode.
+            // swiftSettings: [.unsafeFlags(["-enable-testing"])]
         ),
         .testTarget(
             name: "BlueprintUITests",
@@ -35,6 +37,8 @@ let package = Package(
             name: "BlueprintUICommonControls",
             dependencies: ["BlueprintUI"],
             path: "BlueprintUICommonControls/Sources"
+            // Enable this setting to allow running tests in release mode.
+            // swiftSettings: [.unsafeFlags(["-enable-testing"])]
         ),
         .testTarget(
             name: "BlueprintUICommonControlsTests",